Added XO game
This is the Tic Tac Toe game done by a beginner in Python 3. This is not something that strongly holds Hard algorithm nor GUI. It completely runs using IDLE only.
But SURPRISING thing you may note is the use of only Conditional statements,loops,etc .Hence you may expect it to be highly logical in flow.

As I am a beginner , I have did that only in the level of IDLE. You can run it with IDLE or any other IDE which supports running programs such as PyCharm ,Spyder , Anaconda Navigator.	The instructions to play the game will be displayed whenever you play itself.
Multiplayer game only.

Game Features:-
		Unlimited loop- You can choose when to stop the game.
		Error handled well - If player types wrong input ,or any other errors can be handled by the program.
		Well designed , but not with GUI.You will get a clean look in IDLE itself.